requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
discord.py==2.5.2
docopt==0.6.2
fonttools==4.58.0
frozenlist==1.6.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
multidict==6.4.3
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pipreqs==0.4.13
propcache==0.3.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
urllib3==2.4.0
yarg==0.1.10
yarl==1.20.0